+++ src/views/HomeView.vue
@@ -7,11 +7,11 @@
   <div>
     <h1>YouTube Audio/Video Downloader</h1>
     <div class="input-container">
-      <!-- Input field for entering URLs -->
       <input
         v-model="videoUrl"
         placeholder="Enter YouTube Video URL"
         @keyup.enter="addUrlToList"
+        @contextmenu.prevent="pasteFromClipboard"
       />
       <div class="add-url-button" @click="addUrlToList" title="Add another URL">+</div>
 
@@ -28,7 +28,6 @@
         {{ downloading ? 'Downloading...' : 'Download' }}
       </button>
 
-      <!-- List of URLs to download -->
       <br>
       <div class="url-list-container">
         <div class="url-list">
@@ -40,14 +39,13 @@
       </div>
     </div>
 
-    <!-- Notification component -->
     <NotificationComponent :message="notificationMessage" :type="notificationType" />
   </div>
 </template>
 
 <script>
-import axios from 'axios'; // Import Axios for making HTTP requests
-import NotificationComponent from '../components/NotificationComponent.vue'; // Import the Notification component
+import axios from 'axios'; 
+import NotificationComponent from '../components/NotificationComponent.vue'; 
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { ipcRenderer } from 'electron';
 
@@ -69,6 +67,11 @@ export default {
     };
   },
   methods: {
+    pasteFromClipboard() {
+      navigator.clipboard.readText().then((text) => {
+        this.videoUrl = text; // Paste the clipboard content into the input field
+      });
+    },
     validateUrl() {
       if (this.validUrlPattern.test(this.videoUrl)) {
         return true
@@ -79,39 +82,38 @@ export default {
     async addUrlToList() {
       if (this.videoUrl && this.validateUrl(this.videoUrl)) {
         try {
-          // Make an HTTP GET request to the video URL to fetch the HTML content
           const response = await axios.get(this.videoUrl);
 
           // Extract the video title from the HTML source using a regular expression
-          const titleMatch = response.data.match(/<title[^>]*>([^<]+)<\/title>/i);
-          const videoTitle = titleMatch ? titleMatch[1] : 'Unknown Title';
+          const titleMatch = response.data.match(/<title[^>]*>([^<]+)<\/title>/i)
+          const videoTitle = titleMatch ? titleMatch[1] : 'Unknown Title'
 
-          this.urlList.push({ url: this.videoUrl, title: videoTitle });
-          this.videoUrl = ''; // Clear the input field
+          this.urlList.push({ url: this.videoUrl, title: videoTitle })
+          this.videoUrl = ''
         } catch (error) {
-          console.error('Error fetching video details:', error);
-          this.showNotification('Error fetching video details.', 'error');
+          console.error('Error fetching video details:', error)
+          this.showNotification('Error fetching video details.', 'error')
         }
       } else if (this.videoUrl.length === 0) {
-        this.showNotification('Nothing to add.', 'error');
+        this.showNotification('Nothing to add.', 'error')
       }
     },
     removeUrlFromList(index) {
-      this.urlList.splice(index, 1);
+      this.urlList.splice(index, 1)
     },
     async downloadVideos() {
       if (this.videoUrl.length !== 0) {
         await this.addUrlToList()
       } else if (this.videoUrl.length === 0 && this.urlList.length === 0) {
-        this.showNotification('Please add at least one URL to download.', 'error');
-        return;
+        this.showNotification('Please add at least one URL to download.', 'error')
+        return
       }
 
       // Set downloading to true to show the loading indicator
-      this.downloading = true;
+      this.downloading = true
 
       for (const urlObject of this.urlList) {
-        const url = urlObject.url;
+        const url = urlObject.url
         try {
           // Use IPC to trigger video download in the main process
           ipcRenderer.send('download-video', {
@@ -120,26 +122,26 @@ export default {
             audioOnly: this.downloadAudioOnly, // Pass the download mode
           });
 
-          this.showNotification(`${urlObject.title} downloaded successfully.`, 'success');
+          this.showNotification(`${urlObject.title} downloaded successfully.`, 'success')
         } catch (error) {
-          console.error(`Error downloading ${urlObject.title}:`, error);
-          this.showNotification(`Error downloading ${urlObject.title}.`, 'error');
+          console.error(`Error downloading ${urlObject.title}:`, error)
+          this.showNotification(`Error downloading ${urlObject.title}.`, 'error')
         }
       }
 
       // Clear the list of URLs and reset the download state
-      this.urlList = [];
-      this.downloading = false;
+      this.urlList = []
+      this.downloading = false
     },
     toggleDownloadMode() {
-      this.downloadAudioOnly = !this.downloadAudioOnly;
+      this.downloadAudioOnly = !this.downloadAudioOnly
     },
     showNotification(message, type) {
-      this.notificationMessage = message;
-      this.notificationType = type;
+      this.notificationMessage = message
+      this.notificationType = type
       setTimeout(() => {
-        this.notificationMessage = '';
-        this.notificationType = '';
+        this.notificationMessage = ''
+        this.notificationType = ''
       }, 5000); // Hide the notification after 5 seconds
     },
   },